Restore search-filter bidirectionality
This makes the search-filter bidirectional again by explicitly adding a
watch from the .input.searchFilter back to .searchFilter, where the
bidirectional binding can take it up.

This solves the issue only in one direction (changes to the variable
from outside still won't propagate in), and feels needless, but works
for the purpose of the bug report.

Closes: https://github.com/myforce/angularjs-dropdown-multiselect/issues/19