Add deselectToLimit option and documentation
When the option is set and an item gets selected, old items are removed
from the selected-model until the selectionLimit is reached.

This is especially handy when used with selectionLimit:1 but an Array as
selected-model, as it will then tolerate (and correctly show) models
with several selected items pre-configured while maintaining the
constraint that the user may only select one item.