Drop externalEvents to allow events to be updated
While $scope.events is received from outside the directive using a
binding, it only got evaluated at construction time due to the
externalEvents dictionary, which was presumably created to easily
provide default noop event handlers.

This patch abolishes externalEvents and replaces it with a
$scope._events function, which does a lookup in the scope's events at
runtime, and produces a noop as a last resort.

Thus, applications using this directive can pass in an event handler
object and add or remove handlers in it at any time.